fix mutations

+++ server/schema/resolvers.js
@@ -5,25 +5,44 @@ const { signToken } = require('../utils/auth');
 const resolvers = {
     Query:{
         getAllUsers: async () =>{
-          return await User.find({})
+            return await User.find({})
         },
 
-        user: async (parent, args, context) =>{
-            if (context.user) {
-                const userData = await User.findOne({ _id: context.user._id })
-                    .select('-__v -password')
-                return userData;
+        user: async (parent, args) =>{
+           
+            return await User.findById(args._id);
+        },
+
+        stories: async (parent, {userId, name})=>{
+           
+            const params = {};
+
+            if (userId) {
+                params.userId = userId;
             }
-            throw new AuthenticationError('You need to be logged in!');
+
+            if (name) {
+                params.name = {
+                $regex: name
+                };
+            }
+
+            return await Story.find(params).populate('userId');
+        },
+
+        story: async (parent, args) =>{
+            const story = await Story.findById(args._id).populate('userId');
+            console.log(story);
+            return story; 
         },
         
-        story: async (parent, args) =>{
-            return await Story.findById(args.id);
-        },
+        me: async (parent, args, context) => {
+            if (context.user) {
+              return User.findOne({ _id: context.user._id });
+            }
+            throw new AuthenticationError('You need to be logged in!');
+          },
 
-        stories: async (parent, args)=>{
-            return await Story.find({});
-        }
         },
 
     Mutation:{
@@ -52,42 +71,47 @@ const resolvers = {
         },
         
         // Delete a user
-        deleteUser: async (parent, args) => {
-            try{
+        deleteUser: async (parent, args, context) => {
+            if(context.user){
             Story.find({ userId: args.id }).then((stories) => {
               stories.forEach((story) => {
                 story.remove();
               });
             });
-    
+       
             return await User.findByIdAndRemove(args.id);
-        }catch(err){
-            console.log(err);
         }
         },
        // Add a story
-        addStory: async (parent, args) => {
-            try{
-                const story = new Story({
-                    title: args.title,
-                    description: args.description,
-                    image: args.image,
-                    userId: args.userId,
-                  });
-          
-                  return await story.save();
-            }catch(err){
-                console.log(err);
+        addStory: async (parent, args, context) => {
+         if(context.user){
+            const story = Story.create({
+                title: args.title,
+                description: args.description,
+                image: args.image,
+                userId: args.userId,
+              });
+              await User.findByIdAndUpdate(context.user._id,{
+                  $addToSet: {stories: story}
+              })  
+           return story;
             }
         },
         
         // Delete a story
-        deleteStory: async (parent, args) =>{
-            return await Story.findByIdAndRemove(args.id);
+        deleteStory: async (parent, args, context) =>{
+            if(context.user){
+            const story = Story.findByIdAndRemove(args.id);
+            await User.findByIdAndUpdate(context.user._id,{
+                $pull: {stories: story}
+            }) 
+            return story;
+            }
         },
 
         // Update a story
         updateStory: async (parent, args) => {
+            if(context.user){
             return await Story.findByIdAndUpdate(
                 args.id,
                 {
@@ -99,6 +123,7 @@ const resolvers = {
                 },
                 { new: true }
               );
+            }
         }
     }
 }
